+++ src/aigora/curriculum_graph/application/queries/query_errors.py
